Added token verify in firewall, Refactor operations api and Fixes
