+++ firewall api/controller/controller.go
@@ -3,22 +3,22 @@ package controller
 import (
     "encoding/json"
     "log"
+    "fmt"
     "errors"
-    // "fmt"
-    // "../db"
+    "strings"
     "bytes"
     "../model"
     "io/ioutil"
     "net/http"
 
     "github.com/gorilla/mux"
-    // jwt "github.com/dgrijalva/jwt-go"
+    jwt "github.com/dgrijalva/jwt-go"
     // "go.mongodb.org/mongo-driver/bson"
     // "golang.org/x/crypto/bcrypt"
 )
 
 ////////////////////////////////////////////////////
-// Login api
+// Login api handlers
 ////////////////////////////////////////////////////
 
 func Register(w http.ResponseWriter, r *http.Request) {
@@ -69,8 +69,11 @@ func Register(w http.ResponseWriter, r *http.Request) {
             return
         }
 
+        // Create and send response
         if res.Result == "User created successfully" {
-            responseJSON(w, "Registration successful", "Registration successful")
+            var response model.ResponseResult
+            response.Result = "Registration successful"
+            responseJSON(w, "Registration successful", response)
             return
         }
 
@@ -122,7 +125,10 @@ func Login(w http.ResponseWriter, r *http.Request) {
         log.Println("Add token in DB")
         log.Println(token)
 
-        responseJSON(w, "Login successful", "Login successful")
+        // Create and send response
+        var response model.ResponseResult
+        response.Result = "Login successful"
+        responseJSON(w, "Login successful", response)
         return
     }
 
@@ -130,13 +136,86 @@ func Login(w http.ResponseWriter, r *http.Request) {
     return
 }
 
+func Delete(w http.ResponseWriter, r *http.Request) {
+    log.Println("Received request: delete")
+
+    // Read body
+    params := mux.Vars(r)
+    username := params["username"]
+    // Read body of the request
+
+    // Sending request to login-api
+    response, err := http.Post("http://login-api:5002/delete/" + username, "application/json", nil)
+    if err != nil {
+        log.Println("Error in sending request to login-api")
+        log.Println(err)
+        return
+    }
+
+    // Unmarshall response from login-api
+    var res model.ResponseResult
+    body, _ := ioutil.ReadAll(response.Body)
+    err = json.Unmarshal(body, &res)
+    if err != nil {
+        responseError(w, "Cannot unmarshal response from login-api", err, http.StatusInternalServerError)
+        return
+    }
+
+    if res.Error == "" {
+        if res.Result == "Invalid user" {
+            responseError(w, "Invalid username", errors.New("Invalid username"), http.StatusBadRequest)
+            return
+        }
+
+        // Delete user from data db
+        response, err := http.Post("http://operations-api:5000/user/delete/" + username, "application/json", nil)
+        if err != nil {
+            log.Println("Error in sending request to operations-api")
+            log.Println(err)
+            return
+        }
+
+        // Unmarshall response from operations-api
+        var res model.ResponseResult
+        body, _ := ioutil.ReadAll(response.Body)
+        err = json.Unmarshal(body, &res)
+        if err != nil {
+            responseError(w, "Cannot unmarshal response from operations-api", err, http.StatusInternalServerError)
+            return
+        }
+
+        // Create and send response
+        if res.Result == "Deletion successful" {
+            var response model.ResponseResult
+            response.Result = "Deletion successful"
+            responseJSON(w, "Deletion successful", response)
+            return
+        }
+
+        responseError(w, "Cannot delete user", nil, http.StatusInternalServerError)
+        return
+    }
+
+    responseError(w, "Cannot delete user", nil, http.StatusInternalServerError)
+    return
+}
+
 ////////////////////////////////////////////////////
-// Operations api
+// Operations api handlers
 ////////////////////////////////////////////////////
 func CreateSubscription(w http.ResponseWriter, r *http.Request) {
     log.Println("Received request: create subscription")
 
     // TODO: Check authentication token
+    tokenString := r.Header.Get("Authorization")
+    log.Println(tokenString)
+    tokenString = strings.Split(tokenString, "Bearer ")[1]
+    ok := true
+    // user, ok := verifyToken(tokenString)
+    if !ok {
+        responseError(w, "Token is not valid", nil, http.StatusInternalServerError)
+        return
+    }
 
     // Read body
     params := mux.Vars(r)
@@ -159,9 +238,12 @@ func CreateSubscription(w http.ResponseWriter, r *http.Request) {
         return
     }
 
-
+    // Create and send response
     if res.Error == "" {
-        responseJSON2(w, "Operation successful", "Operation successful", res.Data)
+        var response model.ResponseResult
+        response.Result = "Operation successful"
+        response.Data = res.Data
+        responseJSON(w, "Operation successful", response)
         return
     }
 
@@ -173,6 +255,15 @@ func GetSubscriptions(w http.ResponseWriter, r *http.Request) {
     log.Println("Received request: get all subscriptions")
 
     // TODO: Check authentication token
+    tokenString := r.Header.Get("Authorization")
+    log.Println(tokenString)
+    tokenString = strings.Split(tokenString, "Bearer ")[1]
+    ok := true
+    // user, ok := verifyToken(tokenString)
+    if !ok {
+        responseError(w, "Token is not valid", nil, http.StatusInternalServerError)
+        return
+    }
 
     // Read body
     params := mux.Vars(r)
@@ -195,8 +286,12 @@ func GetSubscriptions(w http.ResponseWriter, r *http.Request) {
         return
     }
 
+    // Create and send response
     if res.Error == "" {
-        responseJSON2(w, "Operation successful", "Operation successful", res.Data)
+        var response model.ResponseResult
+        response.Result = "Operation successful"
+        response.Data = res.Data
+        responseJSON(w, "Operation successful", response)
         return
     }
 
@@ -208,6 +303,15 @@ func GetSubscription(w http.ResponseWriter, r *http.Request) {
     log.Println("Received request: get one subscription")
 
     // TODO: Check authentication token
+    tokenString := r.Header.Get("Authorization")
+    log.Println(tokenString)
+    tokenString = strings.Split(tokenString, "Bearer ")[1]
+    ok := true
+    // user, ok := verifyToken(tokenString)
+    if !ok {
+        responseError(w, "Token is not valid", nil, http.StatusInternalServerError)
+        return
+    }
 
     // Read body
     params := mux.Vars(r)
@@ -231,8 +335,12 @@ func GetSubscription(w http.ResponseWriter, r *http.Request) {
         return
     }
 
+    // Create and send response
     if res.Error == "" {
-        responseJSON2(w, "Operation successful", "Operation successful", res.Data)
+        var response model.ResponseResult
+        response.Result = "Operation successful"
+        response.Data = res.Data
+        responseJSON(w, "Operation successful", response)
         return
     }
 
@@ -249,7 +357,6 @@ func responseError(w http.ResponseWriter, logMessage string, err error, code int
     if err != nil {
         log.Println(err.Error())
     }
-    log.Println("here")
 
     w.Header().Set("Content-Type", "application/json")
     w.WriteHeader(code)
@@ -259,20 +366,41 @@ func responseError(w http.ResponseWriter, logMessage string, err error, code int
     json.NewEncoder(w).Encode(res)
 }
 
-func responseJSON(w http.ResponseWriter, logMessage string, result string) {
-    log.Println(logMessage)
-    w.Header().Set("Content-Type", "application/json")
-    var res model.ResponseResult
-    res.Result = result
-    json.NewEncoder(w).Encode(res)
-}
-
-func responseJSON2(w http.ResponseWriter, logMessage string, result string, data []model.Subscription) {
-    log.Println(logMessage)
+func responseJSON(w http.ResponseWriter, message string, data interface{}) {
+    log.Println(message)
     w.Header().Set("Content-Type", "application/json")
     w.WriteHeader(http.StatusOK)
-    var res model.ResponseResult
-    res.Result = result
-    res.Data = data
-    json.NewEncoder(w).Encode(res)
+    json.NewEncoder(w).Encode(data)
+}
+
+func verifyToken(tokenString string) (model.User, bool){
+    log.Println(tokenString)
+    var result model.User
+
+    token, err := jwt.Parse(tokenString, func(token *jwt.Token) (interface{}, error) {
+        // Don't forget to validate the alg is what you expect:
+        if _, ok := token.Method.(*jwt.SigningMethodHMAC); !ok {
+            return nil, fmt.Errorf("Unexpected signing method")
+        }
+        return []byte("secret"), nil
+    })
+    if err != nil {
+        log.Println("Error in jwt.Parse function")
+        return result, false
+    }
+
+    log.Println("after parsing")
+    // var res model.ResponseResult
+    if claims, ok := token.Claims.(jwt.MapClaims); ok && token.Valid {
+        result.Username = claims["username"].(string)
+        result.FirstName = claims["firstname"].(string)
+        result.LastName = claims["lastname"].(string)
+        log.Println("here")
+        // json.NewEncoder(w).Encode(result)
+        return result, true
+    } else {
+        // res.Error = err.Error()
+        log.Println("here2")
+        return result, false
+    }
 }
