+++ operations api/db/db.go
@@ -1,7 +1,7 @@
 package db
 
 import (
-    "fmt"
+    // "fmt"
     "log"
     "os"
 
@@ -10,9 +10,8 @@ import (
 
 func GetDBCollection() (*mgo.Collection, *mgo.Session, error) {
     // Connect to mongo
-    session, err := mgo.Dial("mongo-credentials:27017")
+    session, err := mgo.Dial("mongo:27017")
     if err != nil {
-        fmt.Println("db eroare")
         log.Fatalln(err)
         log.Fatalln("mongo err")
         os.Exit(1)
@@ -20,7 +19,7 @@ func GetDBCollection() (*mgo.Collection, *mgo.Session, error) {
     session.SetMode(mgo.Monotonic, true)
 
     // Get users collection
-    collection := session.DB("app").C("users")
+    collection := session.DB("app").C("subscriptions")
 
     return collection, session, nil
 }
